RMS-11 RMS-110 RMS-111 Added store, refactored CreateProduct page, json-server can be used as BE server
