+++ .env.local
@@ -1,0 +1,2 @@
+REACT_APP_PROFILE=json-server-api
+REACT_APP_BASE_URL=http://localhost:3000
