+++ package.json
@@ -37,6 +37,7 @@
     "react-notification-alert": "0.0.13",
     "react-perfect-scrollbar": "1.5.8",
     "react-quill": "1.3.5",
+    "react-redux": "^7.2.6",
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
     "react-scripts": "4.0.3",
@@ -44,6 +45,8 @@
     "react-select2-wrapper": "1.0.4-beta6",
     "react-to-print": "2.12.2",
     "reactstrap": "8.9.0",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
     "select2": "4.0.13",
     "sweetalert2": "10.13.0",
     "web-vitals": "^1.1.2",
@@ -82,6 +85,7 @@
     "eslint-plugin-flowtype": "5.2.0",
     "gulp": "4.0.2",
     "gulp-append-prepend": "1.0.8",
+    "redux-devtools-extension": "^2.13.9",
     "typescript": "4.1.3"
   }
 }
