+++ src/http-common.js
@@ -1,8 +1,9 @@
 import axios from "axios";
 
 export const axiosConfig = axios.create({
-  baseURL: "http://localhost:3000",
+  baseURL: process.env.REACT_APP_BASE_URL,
   headers: {
-    "Content-type": "application/json"
+    "Content-type": "application/json",
+    "Accept": "application/json"
   }
 });
