+++ src/redux/root.reducer.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import { productReducer } from "./products/product.reducer";
+
+export const rootReducer = combineReducers({
+  product: productReducer
+});
