+++ src/redux/types.actions.js
@@ -1,0 +1,3 @@
+export const CREATE_PRODUCT_LOADING = "CREATE_PRODUCT_LOADING";
+export const CREATE_PRODUCT_ERROR = "CREATE_PRODUCT_ERROR";
+export const CREATE_PRODUCT_COMPLETE = "CREATE_PRODUCT_COMPLETE";
